requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
httpx
djangorestframework